bugfixing, prep and planning
implemented initial arm angular velocity, arm approach angle, arm approach angular velocity, need to flesh them out and finalize them. fixed a bug where the game would crash because creating an attack couldn't grab the scene from the mid-destruction player before the player was done destroying but after it was gone from the scene. added a flip function to the base entity in prep for making enemies flip and have flip cooldowns. renamed player attack spawning function so this.attack could be a reference to the current attack. added gravity to the base object and a temporary floor.